Backend 3.0 3/VerbX-App/ivish/backend/app/requirements.txt
# ======================
# 🚀 Ivish AI Backend Requirements
# ======================
# Security-pinned versions for production
# Offline-capable and edge-optimized
# Modular, testable, and deployable stack

# 🧠 CORE AI (Quantized/Edge-Optimized)
torch==2.8.0   # Updated to compatible torch version
transformers==4.41.0  # HuggingFace models
optimum==1.18.1  # Quantized inference
onnxruntime==1.23.0  # Edge inference
# whispercpp @ git+https://github.com/ggerganov/whisper.cpp.git@master#subdirectory=python  # Offline Whisper
# sentencepiece==0.1.99  # Tokenization
tokenizers==0.15.2  # Fast text processing
torchaudio==2.8.0   # Audio I/O
torchvision==0.23.0   # Vision support (optional)

# 🗣️ AUDIO PROCESSING (Low-Latency)
soundfile==0.12.1  # Audio I/O
pydub==0.25.1  # Audio manipulation
webrtcvad==2.0.10  # Voice activity detection
ffmpeg-python==0.2.0  # Media handling
librosa==0.10.1  # Feature extraction
numpy==1.26.4  # Numerical operations
scipy==1.13.0

# 🌐 BACKEND & REALTIME
fastapi==0.111.0  # API framework
uvicorn==0.30.0  # ASGI server
websockets==12.0  # Realtime communication
httpx==0.27.0  # Async HTTP
aiofiles==23.2.1  # Async file I/O
msgpack==1.0.8  # Binary serialization
starlette==0.37.2  # FastAPI internal dependency
python-socketio==5.11.2  # Socket.IO server/client
python-engineio==4.9.0  # Engine.IO layer (needed for socketio)
requests==2.32.3  # Extra HTTP support

# 🔐 SECURITY (Military-Grade)
cryptography==42.0.5  # AES/RSA/PBKDF2
pyjwt[crypto]==2.8.0  # JWT tokens
pynacl==1.5.0  # Libsodium bindings
bcrypt==4.1.2  # Password hashing
argon2-cffi==23.1.0  # Memory-hard hashing
python-jose==3.3.0  # JWT/JSON Web Keys
# secure-sandbox==0.1.3  # Secure execution context
# hmac==3.0.1  # Message authentication

# 🛡️ BLOCKCHAIN & ZKP
web3==6.17.0  # Ethereum compatibility
eth-utils>=5.0.0  # Ethereum utilities
pymerkletools==1.0.3  # Merkle proofs
# zkp @ git+https://github.com/ivish-ai/zkp-proofs.git@v1.2.0  # Custom ZKP

# 🗄️ DATABASES & CACHE
redis[hiredis]==5.0.4  # Low-latency cache
pymongo[srv]==4.6.3  # MongoDB driver
motor==3.3.2  # Async MongoDB
tinydb==4.8.0  # Lightweight document store (fallback)

# 📦 UTILITIES
pydantic==2.7.1  # Data validation
loguru==0.7.2  # Structured logging
python-dotenv==1.0.1  # Config mgmt
tqdm==4.66.2  # Progress bars
regex==2024.4.28  # Unicode-aware regex
orjson==3.10.1  # Fast JSON serialization
protobuf==4.21.12  # Efficient data serialization
psutil==5.9.8  # System and process metrics
rich==13.7.0  # CLI rendering

# 🌍 MULTILINGUAL SUPPORT
indic-transliteration==1.6.2  # Indic languages
langdetect==1.0.9  # Language ID
spacy==3.7.4  # NLP pipelines
spacy-lookups-data==1.0.3  # Lexical data
fasttext==0.9.2  # Language detection
textblob==0.17.1  # Sentiment fallback
nltk==3.8.1  # Linguistic analysis
deep-translator==1.11.5  # Language translation
googletrans==4.0.0-rc1  # Translation fallback
transformers_stream_generator==0.0.1  # Streaming for large models

# 🧬 FEDERATED & SELF LEARNING
flower==1.8.0  # Federated learning
syft==0.8.5  # Secure aggregation
joblib==1.4.0  # Model serialization
# autoencoder @ git+https://github.com/ivish-ai/autoencoder.git@v1.0.0  # Auto-learning
# model-validator @ git+https://github.com/ivish-ai/model_validator.git@v1.1.0  # AutoCoder safety

# 🧪 TESTING
pytest==8.2.1  # Test framework
pytest-asyncio==0.23.6  # Async tests
locust==2.27.0  # Load testing
hypothesis==6.103.0  # Property tests
faker==25.8.0  # Test data
coverage==7.5.0  # Code coverage
mypy==1.9.0  # Type checking
pytest-cov==4.1.0  # Coverage reporting
pytest-xdist==3.5.0  # Parallel testing

# 🧠 SELF-LEARNING & AUTO-CODER
# gpt-prompter @ git+https://github.com/ivish-ai/gpt-prompter.git@v1.0.0  # Prompt generation
copier==6.1.0  # Template-based codegen
jinja2==3.1.4  # Codegen templates
black==24.3.0  # Code formatting
flake8==7.1.1  # Linting
pylint==3.0.3  # Code analysis

# 📊 PERFORMANCE & MONITORING
prometheus-client==0.16.0  # Metrics exposure
memory-profiler==0.6.0  # Memory tracking
async-timeout==4.0.3  # Async time control
circuit-breaker==1.1.0  # Resilience pattern
rate-limit==2.3.0  # Rate limiting
pyinstrument==4.5.0  # Performance tracing

# Google Auth
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# 📁 DEVOPS & INFRA
GitPython==3.1.43  # Git integration
docker==6.1.5  # Containerization
ansible==9.4.0  # Configuration
kubernetes==28.1.1  # Orchestration
gunicorn==21.2.0  # Production server
uvloop==0.19.0  # Faster asyncio loop
